requirements.txt
# Core ML/NLP Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
peft>=0.4.0
trl>=0.7.0
bitsandbytes>=0.41.0

# OpenAI API for dataset generation
openai>=0.27.0
python-dotenv>=1.0.0

# Additional utilities
accelerate>=0.20.0
pynvml>=11.4.1
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Logging and monitoring
wandb>=0.15.0

# Development
jupyter>=1.0.0
ipython>=8.14.0
black>=23.7.0
flake8>=6.0.0
